+++ tests/testthat/test-bit64-package.R
@@ -173,7 +173,7 @@ test_that("cbind and rbind", {
   expect_true(identical.integer64(rbind(as.integer64(1:3), 1:3), t(x)))
 })
 
-test_that("Coercion", {
+test_that("Coercion from integer64 works as intended", {
   expect_identical(
     as.double(as.integer64(c(NA, seq(0.0, 9.0, 0.25)))),
     as.double(as.integer(c(NA, seq(0.0, 9.0, 0.25))))
@@ -194,6 +194,9 @@ test_that("Coercion", {
     as.integer(as.integer64(c(NA, FALSE, TRUE))),
     as.integer(c(NA, FALSE, TRUE))
   )
+})
+
+test_that("Coercion to integer64 with as.integer64 works as intended", {
   expect_identical(
     as.integer64(as.integer(as.integer64(-9:9))),
     as.integer64(-9:9)
@@ -203,12 +206,111 @@ test_that("Coercion", {
     as.integer64(-9:9)
   )
   expect_identical(
+    as.integer64(as.complex(-9:9)),
+    as.integer64(-9:9)
+  )
+  expect_identical(
+    as.integer64(as.raw(0:9)),
+    as.integer64(0:9)
+  )
+  expect_identical(
     as.integer64(as.character(as.integer64(-9:9))),
     as.integer64(-9:9)
   )
   expect_identical(
     as.integer64(as.character(lim.integer64())),
     lim.integer64()
+  )
+})
+
+test_that("Coercion to integer64 with as() works as intended", {
+  expect_identical(
+    as(as.raw(1L), "integer64"),
+    as.integer64(1L)
+  )
+  expect_identical(
+    as(TRUE, "integer64"),
+    as.integer64(1L)
+  )
+  expect_identical(
+    as(111L, "integer64"),
+    as.integer64(111L)
+  )
+  expect_identical(
+    as(111.0, "integer64"),
+    as.integer64(111L)
+  )
+  expect_identical(
+    as(111.0 + 0.0i, "integer64"),
+    as.integer64(111L)
+  )
+  expect_identical(
+    as("111", "integer64"),
+    as.integer64(111L)
+  )
+  expect_identical(
+    as(as.factor(111L), "integer64"),
+    as.integer64(1L)
+  )
+  expect_identical(
+    as(as.ordered(111L), "integer64"),
+    as.integer64(1L)
+  )
+})
+
+test_that("Coercion from integer64 with as() works as intended", {
+  expect_identical(
+    as(as.integer64(1L), "logical"),
+    TRUE
+  )
+  expect_identical(
+    as(as.integer64(111L), "integer"),
+    111L
+  )
+  expect_identical(
+    as(as.integer64(111L), "numeric"),
+    111.0
+  )
+  expect_identical(
+    as(as.integer64(111L), "character"),
+    "111"
+  )
+  expect_identical(
+    as(as.integer64(111L), "factor"),
+    as.factor("111")
+  )
+  expect_identical(
+    as(as.integer64(111L), "ordered"),
+    as.ordered("111")
+  )
+})
+
+test_that("Coercion works for cases requiring recent R", {
+  skip_unless_r(">= 4.0.0")
+
+  expect_identical(
+    as.complex(as.integer64(c(NA, seq(0.0, 9.0, 0.25)))),
+    as.complex(as.integer(c(NA, seq(0.0, 9.0, 0.25))))
+  )
+
+  coercion_warning = "out-of-range values treated as 0 in coercion to raw"
+  expect_warning({
+    raw_from_int64 = as.raw(as.integer64(c(NA, seq(0.0, 1.25, 0.25))))
+  }, coercion_warning, fixed=TRUE)
+  expect_warning({
+    raw_from_int32 = as.raw(as.integer(c(NA, seq(0.0, 1.25, 0.25))))
+  }, coercion_warning, fixed=TRUE)
+
+  expect_identical(raw_from_int64, raw_from_int32)
+
+  expect_identical(
+    as(as.integer64(1L), "raw"),
+    as.raw(1L)
+  )
+
+  expect_identical(
+    as(as.integer64(111L), "complex"),
+    111.0 + 0.0i
   )
 })